% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{agrupaDUI}
\alias{agrupaDUI}
\title{Widget Tipos de Agrupamientos comenzando desde Diario}
\usage{
agrupaDUI(id)
}
\arguments{
\item{id}{Es el id del elemento a referenciar}
}
\description{
Esta es una funcion para obtener un widget de shiny que permite usar un selector de agrupamiento para las opciones Diario, Mensual o Anual.
}
\details{
Esta funcion hace parte del set de elementos a usar en el script ui.R para facilitar la creacion de interfaces
graficas usando Shiny. La funcion solamente recibe el parametro \code{id} del elemento para poder ser referenciado en el script server.R,
Para poder ser referenciado en el script server..
}
\examples{
agrupaDUI("agrup")
}
