% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{barPlotly}
\alias{barPlotly}
\title{Plotly Bar Chart}
\usage{
barPlotly(input, output, session, xdata, ydata, xaxis, setElement = NULL)
}
\arguments{
\item{input}{Required Param}

\item{output}{Required Param}

\item{session}{Required Param}

\item{xdata}{Array with numeric values, usually is a percentage}

\item{ydata}{Array with character values, usually is a name}

\item{xaxis}{x Axis Title}

\item{setElement}{Element name to be highlighted}
}
\description{
This function is used to generates a bar chart using Plotly Package
}
\details{
This function belongs to server.R script to easily generates a bar chart using Dygraphs.
This function has three required params \code{input}, \code{output}, \code{session} that always is used by Shiny Modules.
The \code{ydata} param means to names in y axis, \code{xdata} param is realted to x axis values, usually it must a percentage,
the \code{xaxis} param is to set x axis title, the \code{setElement} param is used to set a element name to be highlighted in the chart.
How it used Shiny Modules, The function will be called using the callModule function, in
the first param is the function to be called, the second param is the identifier used in plotlyObjectUI("id") and the other
params are the same after the session param. See the examples please.
}
\examples{
callModule(barPlotly, "plot1", data1$x, data2$y, "Participacion \%", "ISGG")
callModule(barPlotly, "plot2", data1$x, data2$y, "Ingresos \%")
}
