% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{buttonUI}
\alias{buttonUI}
\title{Widget para Botones}
\usage{
buttonUI(id, nombre)
}
\arguments{
\item{id}{Es el id del elemento a referenciar}

\item{nombre}{Es el nombre que tendra el boton}
}
\description{
Esta es una funcion para obtener un widget de shiny que permite crear un boton en Shiny.
}
\details{
Esta funcion hace parte del set de elementos a usar en el script ui.R para facilitar la creacion de interfaces
graficas usando Shiny. La funcion recibe el parametro \code{id} del elemento para poder ser referenciado en el script server.R,
el parametro \code{nombre} se usara para definir el nombre que va a tener el boton
}
\examples{
buttonUI("idGen","Generar")
}
