% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{chartDygraphsUI}
\alias{chartDygraphsUI}
\title{Mostrar Salida Dygraph Chart}
\usage{
chartDygraphsUI(id)
}
\arguments{
\item{id}{Es el id del elemento a referenciar}
}
\description{
Esta es una funcion para mostrar objetos de graficas de Dygraphs.
}
\details{
Esta funcion hace parte del set de elementos a usar en el script ui.R para facilitar la creacion de interfaces,
sirve para mostrar graficas de la libreria dygraphs
}
\examples{
chartDygraphsUI("plot1")
chartDygraphsUI("plot2")
}
