% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{dataTable}
\alias{dataTable}
\title{Generates Data Table}
\usage{
dataTable(input, output, session, data, fileName)
}
\arguments{
\item{input}{Required Argument}

\item{output}{Required Argument}

\item{session}{Required Argument}

\item{data}{DataFrame to ve visualized}

\item{fileName}{File Name when the Data Table is downloaded}
}
\description{
This function is used to generates a Data Table
}
\details{
This function belongs to server.R script to easily generates a bar chart using Dygraphs.
This function has three required params \code{input}, \code{output}, \code{session} that always is used by Shiny Modules.
The \code{data} param is the data frame to ve visualized as a Data Table in the user interface, \code{fileName} param is the name of the file
when the user click the downlaod button.
How it used Shiny Modules, The function will be called using the callModule function, in
the first param is the function to be called, the second param is the identifier used in dataTableUI("id") and the other
params are the same after the session param. See the examples please.
}
\examples{
callModule(dataTable, "table1", data1, "TestFichero1")
callModule(dataTable, "table2", data2, "TestFichero2")
}
