% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{dataTableUI}
\alias{dataTableUI}
\title{Data Table Widget}
\usage{
dataTableUI(id)
}
\arguments{
\item{id}{Element Identifier}
}
\description{
This function is used to generates a Data Table With PDF and xlsx Buttons to Download using DT Package
}
\details{
This function belongs to script ui.R elements to easily creates user interfaces.
This funcion has only one param \code{id} to be referenced in server.R
}
\examples{
dataTableUI("plot1")
dataTableUI("plot2")
}
