% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{dateLastYearUI}
\alias{dateLastYearUI}
\title{Widget Rango de Fecha del Mes Anterior al Actual}
\usage{
dateLastYearUI(id)
}
\arguments{
\item{id}{Es el id del elemento a referenciar}
}
\description{
Esta es una funcion para obtener un widget de shiny que permite usar un combo selector de rango de fechas para el anio anterior al actual.
}
\details{
Esta funcion hace parte del set de elementos a usar en el script ui.R para facilitar la creacion de interfaces
graficas usando Shiny. La funcion solamente recibe el parametro \code{id} del elemento para poder ser referencido en el script server.R,
Para poder ser referenciado en el script server..
}
\examples{
dateLastYearUI("date1")
dateLastYearUI("date2")
}
