% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{dateRangeLastMonthUI}
\alias{dateRangeLastMonthUI}
\title{Date past month Range Widget}
\usage{
dateRangeLastMonthUI(id)
}
\arguments{
\item{id}{Element Identifier}
}
\description{
This function is to obtain an Shiny Widget to generates a Past month Selector.
}
\details{
This function belongs to script UI.R elements to easily creates user interfaces.
This function only have a parameter \code{id} to be referenced.
}
\examples{
dateRangeLastMonthUI("date1")
dateRangeLastMonthUI("date2")
}
