% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{dateTodayUI}
\alias{dateTodayUI}
\title{Date Only a Day Widget}
\usage{
dateTodayUI(id)
}
\arguments{
\item{id}{Element Identifier}
}
\description{
This function is to obtain an Shiny Widget to generates a day Selector by default it has Sys Date.
}
\details{
This function belongs to script UI.R elements to easily creates user interfaces.
This function only have a parameter \code{id} to be referenced in server.R.
}
\examples{
dateTodayUI("date1")
dateTodayUI("date2")
}
