% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{dygraphObjectUI}
\alias{dygraphObjectUI}
\title{Dygraphs Objects}
\usage{
dygraphObjectUI(id)
}
\arguments{
\item{id}{Element Identifier}
}
\description{
This function allows generate Dygraphs Objects e.g time series Charts.
}
\details{
This function belongs to script.R script to easily creates user interfaces.
It allows showing many types of different charts that can be generated using Dygraphs Package.
}
\examples{
dygraphObjectUI("plot1")
dygraphObjectUI("plot2")
}
