% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{loadBarUI}
\alias{loadBarUI}
\title{Loading Bar Widget}
\usage{
loadBarUI()
}
\description{
Esta es una funcion que se activa cuando se esta realizando algun calculo.
}
\details{
Esta funcion hace parte del set de elementos a usar en el script ui.R para facilitar la creacion de interfaces
graficas usando Shiny. La funcion detecta cuando el sistema esta ocupado realizando algun calculo y muestra en pantalla
unas barras azules que indican calculos, desaparecen cuando el sistema se libera
}
\examples{
loadBarUI()
}
