% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{plotlyObjectUI}
\alias{plotlyObjectUI}
\title{Plotly Objects}
\usage{
plotlyObjectUI(id)
}
\arguments{
\item{id}{Element Identifier}
}
\description{
This function is to generates a time series chart using Dygraphs
}
\details{
This function belongs to script.R script to easily creates user interfaces.
It allows showing many types of different charts that can be generated using Plotly Package.
}
\examples{
plotlyObjectUI("plot1")
plotlyObjectUI("plot2")
}
