% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{timeSeriesDygraphs}
\alias{timeSeriesDygraphs}
\title{Grafica T.S Dygraphs}
\usage{
timeSeriesDygraphs(input, output, session, data, titulo, ejey)
}
\arguments{
\item{input}{Argumento Obligatorio}

\item{output}{Argumento Obligatorio}

\item{session}{Argumento Obligatorio}

\item{data}{Objeto del tipo xts con los datos a graficar}

\item{titulo}{Titulo de la grafica}

\item{ejey}{Titulo del eje y}
}
\description{
Esta es una funcion que sirve para generar una grafica de series de tiempo en Dygraphs.
}
\details{
Esta funcion hace parte del set de elementos a usar en el script server.R para facilitar la generacion de graficas usando Shiny.
La funcion tiene tres argumentos obligatorios \code{input}, \code{output}, \code{session}, estos siempre los usan los modulos en Shiny.
El argumento \code{data} tiene que ser del tipo xts, el argumento \code{titulo} corresponde con el titulo principal de la grafica y el argumento \code{ejey}
es para poner el titulo en el eje y de la grafica. Debido a que este es un modulo, su llamado es especial, ver los ejemplos de como se debe llamar esta funcion.
}
\examples{
callModule(timeSeriesDygraphs, "plot1", trmDia_xts, "TRM diaria", "[COP/USD]")
callModule(timeSeriesDygraphs, "plot2", trmDia_xts, "TRM diaria", "[COP/USD]")
}
