% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{timeSeriesDygraphsUI}
\alias{timeSeriesDygraphsUI}
\title{Mostrar Salida Dygraphs Time Series}
\usage{
timeSeriesDygraphsUI(id)
}
\arguments{
\item{id}{Es el id del elemento a referenciar}
}
\description{
Esta es una funcion para mostrar objetos de graficas de Dygraphs.
}
\details{
Esta funcion hace parte del set de elementos a usar en el script ui.R para facilitar la creacion de interfaces,
sirve para generar graficas de Serie de Tiempo usando la libreria dygraphs
}
\examples{
timeSeriesDygraphsUI("plot1")
timeSeriesDygraphsUI("plot2")
}
