% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{timeSeriesPlotly}
\alias{timeSeriesPlotly}
\title{Grafica T.S Plotly}
\usage{
timeSeriesPlotly(input, output, session, datosx, datosy, titulo, ejey,
  suavizada)
}
\arguments{
\item{input}{Argumento Obligatorio}

\item{output}{Argumento Obligatorio}

\item{session}{Argumento Obligatorio}

\item{datosx}{Vector que Corresponde con las Fechas, debe estar en formato POSIXt}

\item{datosy}{Vector que Corresponde con los valores que van en el eje y}

\item{titulo}{Titulo de la grafica}

\item{ejey}{Titulo del eje y}

\item{suavizada}{Si su valor esta en True, debe mostrar una linea suavizada de los datos
sino mostrata los puntos con los valores reales}
}
\description{
Esta es una funcion que sirve para generar una grafica de series de tiempo.
}
\details{
Esta funcion hace parte del set de elementos a usar en el script server.R para facilitar la generacion de graficas usando Shiny.
La funcion tiene tres argumentos obligatorios \code{input}, \code{output}, \code{session}, estos siempre los usan los modulos en Shiny.
El argumento \code{datosx} tiene que ser del tipo POSIXt, \code{datosy} corresponde con los valores del ejey, el argumento \code{titulo}
corresponde con el titulo principal de la grafica y el argumento \code{ejey} es para poner el titulo en el eje y de la grafica.
\code{suavizada} es un parametro que indicasi se mostraran los datos con una linea suavizada o se muestran los datos reales.
Debido a que este es un modulo, su llamado es especial, ver los ejemplos de como se debe llamar esta funcion.
}
\examples{
callModule(timeSeriesPlotly, "plot1", trmDia_xts$x, trmDia_xts$y, "TRM diaria", "[COP/USD]", F)
callModule(timeSeriesPlotly, "plot2", trmDia_xts$x, trmDia_xts$y, "TRM diaria", "[COP/USD]", T)
}
