% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{timeSeriesPlotlyUI}
\alias{timeSeriesPlotlyUI}
\title{Mostrar Salida Plotly Time Series}
\usage{
timeSeriesPlotlyUI(id)
}
\arguments{
\item{id}{Es el id del elemento a referenciar}
}
\description{
Esta es una funcion para mostrar objetos de graficas de Series de Tiempo usando Plotly
}
\details{
Esta funcion hace parte del set de elementos a usar en el script ui.R para facilitar la creacion de interfaces,
sirve para generar una grafica de Serie de Tiempo usando Plotly
}
\examples{
timeSeriesPlotlyUI("plot1")
timeSeriesPlotlyUI("plot2")
}
