% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{addGroupByType}
\alias{addGroupByType}
\title{Agrupar Datos Por Algun Tipo y Sumar los Valores}
\usage{
addGroupByType(data, agrupa, columnaFecha, columnaPeriodo, columnaVal, ...)
}
\arguments{
\item{agrupa}{Tipo de Agrupamiento}

\item{columnaFecha}{Numero de Columna donde esta la Fecha, comenzando desde el 1}

\item{columnaPeriodo}{Numero de Columna donde esta el Periodo, comenzando desde el 1, si los datos no tienen periodo poner valor de cero para no ser tenido en cuenta}

\item{columnaVal}{Numero de Columna donde esta el Valor que se desea Sumar, comenzando desde el 1}

\item{...}{Se deben indicar las Columnas por las cuales se desea realizar el agrupamiento}
}
\description{
Esta funcion sirve para agrupar datos de un dataFrame por algun tipo de agrupamiento y sumar los valores de la columna indicada en uno de sus parametros,
la flexibilidad de esta funcion radica en que permite agrupar por el numero de columnas que el usuario requiera#' @param data Set de Datos
}
\details{
Esta funcion sirve como ayuda a la hora de realizar calculos que requieran agrupar informacion y permite indicar
cual es la columna que se quiere sumar y cuales son las columndas por las cuales se quieren agrupar los datos
}
\examples{
addGroupByType(trmDiaria, "Mensual", 1, 0, 2, "FECHA")
}
