% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{barPlotly}
\alias{barPlotly}
\title{Grafica de Barras en Plotly}
\usage{
barPlotly(input, output, session, xdata, ydata, xaxis, setElement)
}
\arguments{
\item{input}{Argumento Obligatorio}

\item{output}{Argumento Obligatorio}

\item{session}{Argumento Obligatorio}

\item{xdata}{Vector que Corresponde con los valores que van en el eje x, usualmente son valores numericos}

\item{ydata}{Vector que Corresponde con los valores del eje y, usualmente deben ser nombres}

\item{xaxis}{Titulo del eje x}

\item{setElement}{Nombre del elemento que se quiere resaltar en la grafica}
}
\description{
Esta es una funcion que sirve para generar una grafica de barras.
}
\details{
Esta funcion hace parte del set de elementos a usar en el script server.R para facilitar la generacion de graficas usando Shiny.
La funcion tiene tres argumentos obligatorios \code{input}, \code{output}, \code{session}, estos siempre los usan los modulos en Shiny.
El argumento \code{ydata} corresponde a los nombres de lo que sea que queramos comparar, \code{xdata} corresponde con los valores del ejex los cuales
deben ser valores numericos usualmente en porcentaje, el argumento \code{xaxis} corresponde con el titulo del eje x se muestran los datos reales.
\code{setElement} Sirve para setear un elemento que se quiera resaltar, en caso de no querer resaltar ninguno no enviar nada en este argumento.
Debido a que este es un modulo, su llamado es especial, ver los ejemplos de como se debe llamar esta funcion.
}
\examples{
callModule(barPlotly, "plot1", data1$x, data2$y, "Participacion \%", "ISGG")
callModule(barPlotly, "plot2", data1$x, data2$y, "Ingresos \%")
}
