% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{buttonUI}
\alias{buttonUI}
\title{Button Widget}
\usage{
buttonUI(id, name)
}
\arguments{
\item{id}{Element Identifier}

\item{name}{Button Name}
}
\description{
This function is to obtain an Shiny Widget to generates a Button to makes any action.
}
\details{
This function belongs to script UI.R elements to easily creates user interfaces.
This funcion has two params, \code{id} to be referenced in server.R and
\code{name} to show in user interface the button name.
}
\examples{
buttonUI("idGen","Generar")
}
