% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModules.R
\name{loadBarUI}
\alias{loadBarUI}
\title{Loading Bar Widget}
\usage{
loadBarUI()
}
\description{
This function is used to generates a bussy indicator in the user interface when the app is doing any calculation
}
\details{
This function belongs to script ui.R elements to easily creates user interfaces.
It detects when the System is bussy doing any calculation and shows a busy indicator in the user screen, it stop when the
calculation is finished
}
\examples{
loadBarUI()
}
